Updated Project

+++ 24markets-backend/src/main/java/com/markets/dto/OrderDTO.java
@@ -1,5 +1,7 @@
 package com.markets.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,8 +15,8 @@ public class OrderDTO {
     private Long userId;
     private Long marketId;
     private String orderType;
-    private Double quantity;
-    private Double pricePerUnit;
-    private Double totalAmount;
+    private BigDecimal quantity;
+    private BigDecimal pricePerUnit;
+    private BigDecimal totalAmount;
     private String status;
 }
